refactor: bespoke export for client native prover / PXE server store lazy load (#13783)
Fixes: https://github.com/AztecProtocol/aztec-packages/issues/13656 (or
more like avoids the problematic import)

Creates specific (and more descriptive) exports for `bb-prover`,
focusing on where the code should run (client/server) rather than the
task at hand (prover/verifier). This mimics the behavior of other
packages with similar issues.

We still have to figure out publishing of native packages.
