+++ yarn-project/bb-prover/src/prover/client/wasm/bundle.ts
@@ -5,7 +5,7 @@ import type { SimulationProvider } from '@aztec/simulator/client';
 import { BBWASMPrivateKernelProver } from './bb_wasm_private_kernel_prover.js';
 
 export class BBWASMBundlePrivateKernelProver extends BBWASMPrivateKernelProver {
-  constructor(simulationProvider: SimulationProvider, threads = 1, log = createLogger('bb-prover:wasm:bundle')) {
+  constructor(simulationProvider: SimulationProvider, threads = 16, log = createLogger('bb-prover:wasm:bundle')) {
     super(new BundleArtifactProvider(), simulationProvider, threads, log);
   }
 }
