+++ yarn-project/bb-prover/src/prover/client/wasm/lazy.ts
@@ -5,7 +5,7 @@ import type { SimulationProvider } from '@aztec/simulator/client';
 import { BBWASMPrivateKernelProver } from './bb_wasm_private_kernel_prover.js';
 
 export class BBWASMLazyPrivateKernelProver extends BBWASMPrivateKernelProver {
-  constructor(simulationProvider: SimulationProvider, threads = 1, log = createLogger('bb-prover:wasm:lazy')) {
+  constructor(simulationProvider: SimulationProvider, threads = 16, log = createLogger('bb-prover:wasm:lazy')) {
     super(new LazyArtifactProvider(), simulationProvider, threads, log);
   }
 }
