+++ yarn-project/bb-prover/src/prover/index.ts
@@ -1,3 +1,2 @@
-export * from './bb_prover.js';
-export * from './bb_native_private_kernel_prover.js';
+export * from './server/bb_prover.js';
 export * from './proof_utils.js';
