+++ yarn-project/pxe/src/entrypoints/server/utils.ts
@@ -1,8 +1,8 @@
-import { BBNativePrivateKernelProver } from '@aztec/bb-prover';
-import { BBWASMBundlePrivateKernelProver } from '@aztec/bb-prover/wasm/bundle';
+import { BBNativePrivateKernelProver } from '@aztec/bb-prover/client/native';
+import { BBWASMBundlePrivateKernelProver } from '@aztec/bb-prover/client/wasm/bundle';
 import { randomBytes } from '@aztec/foundation/crypto';
 import { createLogger } from '@aztec/foundation/log';
-import { createStore } from '@aztec/kv-store/lmdb-v2';
+import type { AztecAsyncKVStore } from '@aztec/kv-store';
 import { BundledProtocolContractsProvider } from '@aztec/protocol-contracts/providers/bundle';
 import { type SimulationProvider, WASMSimulator } from '@aztec/simulator/client';
 import { SimulationProviderRecorderWrapper } from '@aztec/simulator/testing';
@@ -24,10 +24,11 @@ export function createPXEService(
   aztecNode: AztecNode,
   config: PXEServiceConfig,
   useLogSuffix: string | boolean | undefined = undefined,
+  store?: AztecAsyncKVStore,
 ) {
   const simulationProvider = new WASMSimulator();
   const simulationProviderWithRecorder = new SimulationProviderRecorderWrapper(simulationProvider);
-  return createPXEServiceWithSimulationProvider(aztecNode, simulationProviderWithRecorder, config, useLogSuffix);
+  return createPXEServiceWithSimulationProvider(aztecNode, simulationProviderWithRecorder, config, useLogSuffix, store);
 }
 
 /**
@@ -44,6 +45,7 @@ export async function createPXEServiceWithSimulationProvider(
   simulationProvider: SimulationProvider,
   config: PXEServiceConfig,
   useLogSuffix: string | boolean | undefined = undefined,
+  store?: AztecAsyncKVStore,
 ) {
   const logSuffix =
     typeof useLogSuffix === 'boolean' ? (useLogSuffix ? randomBytes(3).toString('hex') : undefined) : useLogSuffix;
@@ -55,12 +57,12 @@ export async function createPXEServiceWithSimulationProvider(
     l2BlockBatchSize: 200,
   } as PXEServiceConfig;
 
-  const store = await createStore(
-    'pxe_data',
-    PXE_DATA_SCHEMA_VERSION,
-    configWithContracts,
-    createLogger('pxe:data:lmdb'),
-  );
+  if (!store) {
+    // TODO once https://github.com/AztecProtocol/aztec-packages/issues/13656 is fixed, we can remove this and always
+    // import the lmdb-v2 version
+    const { createStore } = await import('@aztec/kv-store/lmdb-v2');
+    store = await createStore('pxe_data', PXE_DATA_SCHEMA_VERSION, configWithContracts, createLogger('pxe:data:lmdb'));
+  }
 
   const prover = await createProver(config, simulationProvider, logSuffix);
   const protocolContractsProvider = new BundledProtocolContractsProvider();
